Update recipe-build.yml
This workflow still does the same thing as the previous workflow that was written. It builds a Docker image from a Dockerfile and pushes it to GHCR. Also, it does not use any third-party actions. it is less likely to break due to changes in the actions. It also installs the `docker-buildx` CLI plugin, which allows anyone to build multiple images.